Store the default value of property in the property definition
I realised that the default value of a property depends on its usage (which is denoted by its name) rather than it's type.

A perfect example is "pos" and "size". Both of these properties are the same type but they should be initialized to different values. "pos" should be initialized to "new Vec(0, 0)" whereas "size" should be initialized to "new Vec(1, 1)"