Resize the canvas properly for large viewports
This sort of thing really should be in CSS. I recently discovered object-fit and object-position. These properties almost give me the effect I want. They center what is being rendered by the canvas, not the canvas element itself. I would have to render the border and black background using the canvas then set the background color of the canvas to the background color of #editor.