Implement the Save button
TL;DR I might port this thing to TypeScript

During this, I found a few bugs around the place. Many of which, are only possible in dynamic languages. I think this project might just be a little too big for JavaScript. The reason I chose to create a web app was because I can throw together a decent UI very quickly.

I'm thinking about porting this to a statically typed language that compiles into JavaScript. TypeScript seems to be the popular choice here but I have found another alternative: JS++. This language is similar to C++ and looks a bit better than TypeScript. It also doesn't need a type definition file for jQuery. JS++ provides plugins for many IDEs but not WebStorm.

For this reason, I'm going to port the level editor to TypeScript.