Finish level parser
I was able to load level 6 with the level editor. Level 6 has a few array comprehensions and variables and the level editor was still able to parse it. One problem was the "waiting time" parameter of "MovingPlatform". The "waiting time" parameter can be a float or the string "piston".

I need to find a better alternative to "waiting time".
  -  At the current font scale, "waiting time" doesn't fit in the .key_cell or the #properties list
  -  "waiting time" is and bad name

I could add a bool parameter to moving platform. So instead of saying
  "waiting time": "piston"
I could say
  piston: true